tweak button actions
