+++ RevenueCatUI/CustomerCenter/Views/ManageSubscriptions/ManageSubscriptionsButtonsView.swift
@@ -21,48 +21,41 @@ import SwiftUI
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
-struct ManageSubscriptionsButtonsView: View {
+struct ManageSubscriptionsButtonsView<Content: View>: View {
 
-    @ObservedObject
-    var viewModel: ManageSubscriptionsViewModel
+    var relevantPathsForPurchase: [CustomerCenterConfigData.HelpPath]
+    let determineFlowForPath: (CustomerCenterConfigData.HelpPath) async -> Void
+    @ViewBuilder let label: (CustomerCenterConfigData.HelpPath) -> Content
 
     var body: some View {
-        ForEach(self.viewModel.relevantPathsForPurchase, id: \.id) { path in
+        ForEach(relevantPathsForPurchase, id: \.id) { path in
             ManageSubscriptionButton(
                 path: path,
-                viewModel: self.viewModel
+                determineFlowForPath: determineFlowForPath,
+                label: label
             )
         }
     }
-
 }
 
 @available(iOS 15.0, *)
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
-private struct ManageSubscriptionButton: View {
+private struct ManageSubscriptionButton<Content: View>: View {
 
     let path: CustomerCenterConfigData.HelpPath
-
-    @ObservedObject
-    var viewModel: ManageSubscriptionsViewModel
+    let determineFlowForPath: (CustomerCenterConfigData.HelpPath) async -> Void
+    let label: (CustomerCenterConfigData.HelpPath) -> Content
 
     var body: some View {
         AsyncButton(action: {
-            await self.viewModel.determineFlow(
-                for: path,
-                activeProductId: viewModel.purchaseInformation?.productIdentifier)
+            await determineFlowForPath(path)
         }, label: {
-            if self.viewModel.loadingPath?.id == path.id {
-                TintedProgressView()
-            } else {
-                Text(path.title)
-            }
+            label(path)
         })
-        .disabled(self.viewModel.loadingPath != nil)
+//        .disabled(self.viewModel.loadingPath != nil)
     }
-
 }
 
 #endif
